The rule_packs in the EC config are ECRulePacks! :-D
Apart from a few tiny mechanical follow-up fixes, this mainly involves
moving types & classes around.
