+++ cmk/ec/main.py
@@ -27,7 +27,7 @@ import sys
 import threading
 import time
 import traceback
-from collections.abc import Callable, Iterable, Iterator, Mapping
+from collections.abc import Callable, Iterable, Iterator, Mapping, Sequence
 from logging import getLogger, Logger
 from pathlib import Path
 from types import FrameType
@@ -51,7 +51,15 @@ from cmk.utils.site import omd_site
 from cmk.utils.type_defs import HostName, TimeperiodName, Timestamp
 
 from .actions import do_event_action, do_event_actions, do_notify, event_has_opened
-from .config import Config, ConfigFromWATO, MatchGroups, Rule, TextMatchResult, TextPattern
+from .config import (
+    Config,
+    ConfigFromWATO,
+    ECRulePack,
+    MatchGroups,
+    Rule,
+    TextMatchResult,
+    TextPattern,
+)
 from .core_queries import HostInfo, query_hosts_scheduled_downtime_depth, query_timeperiods_in
 from .crash_reporting import CrashReportStore, ECCrashReport
 from .event import create_event_from_line, Event
@@ -1179,7 +1187,7 @@ class EventServer(ECServerThread):
 
     # Precompile regular expressions and similar stuff.
     def compile_rules(  # pylint: disable=too-many-branches
-        self, rule_packs: Iterable[dict[str, Any]]
+        self, rule_packs: Sequence[ECRulePack]
     ) -> None:
         self._rules = []
         self._rule_by_id = {}
